+++ docs/tasks/artifacts/sv_corpus_grad/lrm_annex_a_gap/worklist_probes/w1_array_method_call.sv
@@ -1,0 +1,8 @@
+module m;
+  int q[$], r[$];
+  initial begin
+    r = q.find with (item > 3);
+    r = q.find_index with (item == 2);
+    r = q.sum with (item * 2);
+  end
+endmodule