+++ docs/tasks/artifacts/sv_corpus_grad/lrm_annex_a_gap/worklist_probes/w2_built_in_data_type.sv
@@ -1,0 +1,5 @@
+module m;
+  std::process p;
+  std::semaphore s;
+  std::mailbox mb;
+endmodule