+++ docs/tasks/artifacts/sv_corpus_grad/lrm_annex_a_gap/worklist_probes/w3_built_in_function_call.sv
@@ -1,0 +1,4 @@
+module m;
+  std::process p;
+  initial p = std::process::self();
+endmodule