+++ docs/tasks/artifacts/sv_corpus_grad/lrm_annex_a_gap/worklist_probes/w4_upward_name_reference.sv
@@ -1,0 +1,3 @@
+module m;
+  initial $display(top_mod.some_net);
+endmodule