+++ docs/tasks/artifacts/sv_corpus_grad/lrm_annex_a_gap/worklist_probes/w6_solve_before_primary.sv
@@ -1,0 +1,4 @@
+class C;
+  rand int a, b;
+  constraint c1 { solve a before b; }
+endclass