+++ docs/tasks/artifacts/sv_corpus_grad/lrm_annex_a_gap/worklist_probes/w7_inline_constraint_decl.sv
@@ -1,0 +1,5 @@
+class C; rand int a; endclass
+module m;
+  C c;
+  initial if (c.randomize(a) with { a < 5; }) $display("ok");
+endmodule